adding Statistics

+++ database/migrations/2022_02_28_110516_create_proposals_table.php
@@ -47,7 +47,7 @@ class CreateProposalsTable extends Migration
             $table->json('files')->nullable();
 
             $table->decimal('bonus', 10)->nullable();
-            $table->integer('commission')->nullable();
+            $table->decimal('commission')->nullable();
             $table->timestamps();
         });
     }
